# Conflicts:
#	apphub-service/apphub-service-skyxplore/apphub-service-skyxplore-game/src/main/java/com/github/saphyra/apphub/service/skyxplore/game/simulation/process/impl/production_order/ProductionOrderProcess.java

+++ apphub-service/apphub-service-custom/apphub-service-custom-elite_base/src/main/java/com/github/saphyra/apphub/service/custom/elite_base/message_processing/saver/LoadoutSaver.java
@@ -2,6 +2,7 @@ package com.github.saphyra.apphub.service.custom.elite_base.message_processing.s
 
 import com.github.saphyra.apphub.api.admin_panel.model.model.performance_reporting.PerformanceReportingTopic;
 import com.github.saphyra.apphub.lib.performance_reporting.PerformanceReporter;
+import com.github.saphyra.apphub.service.custom.elite_base.common.MessageProcessingDelayedException;
 import com.github.saphyra.apphub.service.custom.elite_base.common.PerformanceReportingKey;
 import com.github.saphyra.apphub.service.custom.elite_base.dao.commodity.CommodityLocation;
 import com.github.saphyra.apphub.service.custom.elite_base.dao.last_update.LastUpdateDao;
@@ -10,7 +11,10 @@ import com.github.saphyra.apphub.service.custom.elite_base.dao.loadout.Loadout;
 import com.github.saphyra.apphub.service.custom.elite_base.dao.loadout.LoadoutDao;
 import com.github.saphyra.apphub.service.custom.elite_base.dao.loadout.LoadoutFactory;
 import com.github.saphyra.apphub.service.custom.elite_base.dao.loadout.LoadoutType;
+import com.google.common.util.concurrent.Striped;
+import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
@@ -18,7 +22,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.Lock;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -28,7 +33,7 @@ import static java.util.Objects.isNull;
 @RequiredArgsConstructor
 @Slf4j
 public class LoadoutSaver {
-    private static final Map<LockKey, Object> LOCKS = new ConcurrentHashMap<>();
+    private static final Striped<Lock> LOCKS = Striped.lock(8);
 
     private final LastUpdateDao lastUpdateDao;
     private final LastUpdateFactory lastUpdateFactory;
@@ -36,46 +41,48 @@ public class LoadoutSaver {
     private final LoadoutFactory loadoutFactory;
     private final PerformanceReporter performanceReporter;
 
+    @SneakyThrows
     public void save(LocalDateTime timestamp, LoadoutType type, CommodityLocation commodityLocation, UUID externalReference, Long marketId, List<String> items) {
         if ((isNull(commodityLocation) || isNull(externalReference)) && isNull(marketId)) {
             throw new IllegalArgumentException("Both commodityLocation or externalReference and marketId is null");
         }
 
         LockKey key = new LockKey(externalReference, type);
-        Object lock = LOCKS.computeIfAbsent(key, lockKey -> new Object());
-        synchronized (lock) {
-            try {
-                lastUpdateDao.save(lastUpdateFactory.create(externalReference, type, timestamp));
+        Lock lock = LOCKS.get(key);
+        if (!lock.tryLock(30, TimeUnit.SECONDS)) {
+            throw new MessageProcessingDelayedException("Lock acquisition failed in class " + getClass().getSimpleName());
+        }
+        try {
+            lastUpdateDao.save(lastUpdateFactory.create(externalReference, type, timestamp));
 
-                Map<String, Loadout> existingLoadouts = getExistingLoadouts(type, externalReference, marketId)
-                    .stream()
-                    .collect(Collectors.toMap(Loadout::getName, Function.identity()));
+            Map<String, Loadout> existingLoadouts = getExistingLoadouts(type, externalReference, marketId)
+                .stream()
+                .collect(Collectors.toMap(Loadout::getName, Function.identity()));
 
-                List<Loadout> newItems = items.stream()
-                    .filter(item -> !existingLoadouts.containsKey(item))
-                    .map(name -> loadoutFactory.create(timestamp, type, commodityLocation, externalReference, marketId, name))
-                    .toList();
+            List<Loadout> newItems = items.stream()
+                .filter(item -> !existingLoadouts.containsKey(item))
+                .map(name -> loadoutFactory.create(timestamp, type, commodityLocation, externalReference, marketId, name))
+                .toList();
 
-                List<String> deletedItems = existingLoadouts.values()
-                    .stream()
-                    .map(Loadout::getName)
-                    .filter(name -> !items.contains(name))
-                    .toList();
+            List<String> deletedItems = existingLoadouts.values()
+                .stream()
+                .map(Loadout::getName)
+                .filter(name -> !items.contains(name))
+                .toList();
 
-                performanceReporter.wrap(
-                    () -> loadoutDao.deleteByExternalReferenceAndLoadoutTypeAndNameIn(externalReference, type, deletedItems),
-                    PerformanceReportingTopic.ELITE_BASE_MESSAGE_PROCESSING,
-                    PerformanceReportingKey.SAVE_LOADOUT_DELETE_ALL.name()
-                );
+            performanceReporter.wrap(
+                () -> loadoutDao.deleteByExternalReferenceAndLoadoutTypeAndNameIn(externalReference, type, deletedItems),
+                PerformanceReportingTopic.ELITE_BASE_MESSAGE_PROCESSING,
+                PerformanceReportingKey.SAVE_LOADOUT_DELETE_ALL.name()
+            );
 
-                performanceReporter.wrap(
-                    () -> loadoutDao.saveAll(newItems),
-                    PerformanceReportingTopic.ELITE_BASE_MESSAGE_PROCESSING,
-                    PerformanceReportingKey.SAVE_LOADOUT_SAVE_ALL.name()
-                );
-            } finally {
-                LOCKS.compute(key, (k, v) -> v == lock ? null : v);
-            }
+            performanceReporter.wrap(
+                () -> loadoutDao.saveAll(newItems),
+                PerformanceReportingTopic.ELITE_BASE_MESSAGE_PROCESSING,
+                PerformanceReportingKey.SAVE_LOADOUT_SAVE_ALL.name()
+            );
+        } finally {
+            lock.unlock();
         }
     }
 
@@ -88,6 +95,9 @@ public class LoadoutSaver {
     }
 
 
-    private record LockKey(UUID externalReference, LoadoutType loadoutType) {
+    @Data
+    private static class LockKey {
+        private final UUID externalReference;
+        private final LoadoutType loadoutType;
     }
 }
